Replace std::num by inherent methods

+++ src/main.rs
@@ -24,7 +24,6 @@ use rustc_serialize::json::{ToJson,Json};
 use std::collections::HashMap;
 use std::env;
 use std::net::Ipv4Addr;
-use std::num::Int;
 use std::sync::Arc;
 use std::sync::atomic::Ordering::Relaxed;
 use std::sync::atomic::{AtomicBool, ATOMIC_BOOL_INIT};
@@ -52,7 +51,7 @@ fn send_tcp_synack(sk: cookie::SipHashKey, chan: &Sender<EmptyTcpPacket>,
     sa_packet.tcphdr.th_sport = s_tcphdr.th_dport;
     sa_packet.tcphdr.th_dport = s_tcphdr.th_sport;
     sa_packet.tcphdr.th_flags = TH_SYN | TH_ACK;
-    sa_packet.tcphdr.th_ack = (Int::from_be(s_tcphdr.th_seq) + 1u32).to_be();
+    sa_packet.tcphdr.th_ack = (u32::from_be(s_tcphdr.th_seq) + 1u32).to_be();
     sa_packet.tcphdr.th_seq =
         cookie::tcp(sa_packet.iphdr.ip_src, sa_packet.iphdr.ip_dst,
                     sa_packet.tcphdr.th_sport, sa_packet.tcphdr.th_dport,
@@ -94,13 +93,13 @@ fn log_tcp_ack(zmq_ctx: &mut zmq::Socket, sk: cookie::SipHashKey,
     let ack_cookie = cookie::tcp(s_iphdr.ip_dst, s_iphdr.ip_src,
                                  s_tcphdr.th_dport, s_tcphdr.th_sport,
                                  sk, ts);
-    let wanted_cookie = (Int::from_be(ack_cookie) + 1u32).to_be();
+    let wanted_cookie = (u32::from_be(ack_cookie) + 1u32).to_be();
     if s_tcphdr.th_ack != wanted_cookie {
         let ts_alt = ts - 0x40;
         let ack_cookie_alt = cookie::tcp(s_iphdr.ip_dst, s_iphdr.ip_src,
                                          s_tcphdr.th_dport, s_tcphdr.th_sport,
                                          sk, ts_alt);
-        let wanted_cookie_alt = (Int::from_be(ack_cookie_alt) + 1u32).to_be();
+        let wanted_cookie_alt = (u32::from_be(ack_cookie_alt) + 1u32).to_be();
         if s_tcphdr.th_ack != wanted_cookie_alt {
             return false;
         }
@@ -108,7 +107,7 @@ fn log_tcp_ack(zmq_ctx: &mut zmq::Socket, sk: cookie::SipHashKey,
     let tcp_data_str =
         String::from_utf8_lossy(&dissector.tcp_data).into_owned();
     let ip_src = s_iphdr.ip_src;
-    let dport = Int::from_be(s_tcphdr.th_dport);
+    let dport = u16::from_be(s_tcphdr.th_dport);
     let mut record: HashMap<String, Json> = HashMap::with_capacity(4);
     record.insert("ts".to_string(), Json::U64(ts));
     record.insert("ip_src".to_string(), Json::String(format!("{}.{}.{}.{}",
